+++ service/fs/mem/mem.go
@@ -1,3 +1,3 @@
-// Package mem implements spec.FileSystem and provides an in-memory file system
+// Package mem implements spec.FS and provides an in-memory file system
 // implementation for abstraction and testing reasons.
 package mem
