+++ service/spec/service.go
@@ -6,7 +6,7 @@ package spec
 type ServiceCollection interface {
 	// FS returns a file system service. It is used to operate on file system
 	// abstractions of a certain type.
-	FS() FileSystem
+	FS() FS
 
 	// ID returns an ID service. It is used to create IDs of a certain type.
 	ID() ID
